2020/10/20201005_hamming_distance/requirements.txt
Distance~=0.1.3
ImageHash~=4.1.0
Pillow~=7.2.0